Split gui components into lightweight and heavyweight ones

+++ Backends/LWJGL3Backend/src/main/java/org/tobi29/scapes/engine/backends/lwjgl3/STBGlyphRenderer.java
@@ -38,7 +38,7 @@ public class STBGlyphRenderer implements GlyphRenderer {
     private final STBTTFontinfo info;
     private final int tiles, pageTiles, pageTileBits, pageTileMask, glyphSize,
             imageSize;
-    private final double tileSize, size, scale;
+    private final float tileSize, size, scale;
     private final ByteBuffer glyphBuffer;
     private final IntBuffer intBuffer1 = BufferUtils.createIntBuffer(1),
             intBuffer2 = BufferUtils.createIntBuffer(1), intBuffer3 =
@@ -102,7 +102,7 @@ public class STBGlyphRenderer implements GlyphRenderer {
     @Override
     public synchronized GlyphPage page(int id) {
         ByteBuffer buffer = container.allocate(imageSize * imageSize << 2);
-        double[] width = new double[pageTiles];
+        float[] width = new float[pageTiles];
         int i = 0;
         int offset = id << pageTileBits;
         for (int y = 0; y < tiles; y++) {
@@ -112,13 +112,13 @@ public class STBGlyphRenderer implements GlyphRenderer {
                 char c = (char) (i + offset);
                 STBTruetype.stbtt_GetCodepointHMetrics(info, c, intBuffer1,
                         intBuffer2);
-                double widthX = intBuffer1.get(0) * scale;
+                float widthX = intBuffer1.get(0) * scale;
                 if (!Character.isISOControl(c)) {
                     STBTruetype.stbtt_GetCodepointBox(info, c, intBuffer1,
                             intBuffer2, intBuffer3, intBuffer4);
-                    double offsetX = FastMath.max(
+                    float offsetX = FastMath.max(
                             size * 0.25f + intBuffer1.get(0) * scale, 0.0f);
-                    double offsetY = FastMath.max(
+                    float offsetY = FastMath.max(
                             size * 1.5f - intBuffer4.get(0) * scale, 0.0f);
                     int renderX = FastMath.max(FastMath.round(offsetX), 0);
                     int renderY = FastMath.max(FastMath.round(offsetY), 0);
@@ -127,8 +127,7 @@ public class STBGlyphRenderer implements GlyphRenderer {
                     renderX += xx;
                     renderY += yy;
                     STBTruetype.stbtt_MakeCodepointBitmap(info, glyphBuffer,
-                            glyphSize, glyphSize, glyphSize, (float) scale,
-                            (float) scale, c);
+                            glyphSize, glyphSize, glyphSize, scale, scale, c);
                     for (int yyy = 0; yyy < sizeY; yyy++) {
                         buffer.position(
                                 (renderY + yyy) * imageSize + renderX << 2);
